[맘스터치] - [매출분석] - [월별매출현황(채널별)] [맘스터치] - [매출분석] - [일별시간대] [맘스터치] - [매출분석] - [대비기간별매출] [맘스터치] - [매출분석] - [매출분석(사업전략팀)] [맘스터치] - [매출분석] - [일별매출현황(포스별)] [맘스터치] - [매출분석] - [월별매출현황(포스별)] [맘스터치] - [매출분석] - [월별매출현황(포스별-영수건수)] 1. 매장(멀티선택매장) 조회조건 검색 로직 수정

+++ src/main/java/kr/co/solbipos/sale/anals/saleAnalsMomsBst/service/impl/SaleAnalsMomsBstMapper.java
@@ -30,5 +30,5 @@ public interface SaleAnalsMomsBstMapper {
     List<DefaultMap<String>> getSaleAnalsMomsBstList(SaleAnalsMomsBstVO saleAnalsMomsBstVO);
 
     /** 해당 날짜의 총매출액 조회 */
-    int getRealSaleAmt(SaleAnalsMomsBstVO saleAnalsMomsBstVO);
+    long getRealSaleAmt(SaleAnalsMomsBstVO saleAnalsMomsBstVO);
 }
